feat: implement LM Studio and OpenAI client integration with model listing and chat functionality
