+++ backend/services/llm/__init__.py
@@ -1,0 +1,8 @@
+from .registry import LLMModelInfo, list_models
+from .service import ask_about_listings
+
+__all__ = [
+    "ask_about_listings",
+    "LLMModelInfo",
+    "list_models",
+]
